zadaniesw/requirements.txt
Django==1.11.1
python-dateutil==2.6.0

